+++ tester/0_repl/src/finalize.c
@@ -6,15 +6,13 @@
 /*   By: gsmereka <gsmereka@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/22 23:16:42 by gsmereka          #+#    #+#             */
-/*   Updated: 2023/01/25 14:59:37 by gsmereka         ###   ########.fr       */
+/*   Updated: 2023/01/27 22:35:09 by gsmereka         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../headers/tester.h"
 
 static int	close_files(t_data *data);
-static void	free_list(void **list);
-static void	free_program_memory(t_data *data);
 
 void	finalize(t_data *data)
 {
@@ -60,46 +58,4 @@ static int	close_files(t_data *data)
 		i++;
 	}
 	return (0);
-}
-
-static void	free_program_memory(t_data *data)
-{
-	if (data->valgrind_path)
-		free(data->valgrind_path);
-	if (data->paths)
-		free_list((void **)data->paths);
-	if (data->process.pid)
-		free(data->process.pid);
-	if (data->process.status)
-		free(data->process.status);
-	if (data->input_tests_fd)
-		free(data->input_tests_fd);
-	if (data->expected_outputs_fd)
-		free(data->expected_outputs_fd);
-	if (data->user_outputs_fd)
-		free(data->user_outputs_fd);
-	if (data->user_error_fd)
-		free(data->user_error_fd);
-	if (data->input_tests_name)
-		free_list((void **)data->input_tests_name);
-	if (data->expected_outputs_name)
-		free_list((void **)data->expected_outputs_name);
-	if (data->user_outputs_name)
-		free_list((void **)data->user_outputs_name);
-	if (data->user_error_name)
-		free_list((void **)data->user_error_name);
-}
-
-static void	free_list(void **list)
-{
-	int	i;
-
-	i = 0;
-	while (list[i])
-	{
-		if (list[i])
-			free(list[i]);
-		i++;
-	}
-	free (list);
 }
